Safari-609.1.11.1/LayoutTests/editing/pasteboard/entries-api/datatransfer-items-drop-directoryReader-expected.txt

Basic test coverage for fileSystemDirectoryReader.readEntries()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS dataTransfer.items.length is 1
PASS entry.isDirectory is true
PASS entry.name is "testFiles"
PASS entry.fullPath is "/testFiles"
PASS childEntries.length is 4
PASS childEntries[0].name is "file1.txt"
PASS childEntries[0].fullPath is "/testFiles/file1.txt"
PASS childEntries[0].isFile is true
PASS childEntries[1].name is "file2.txt"
PASS childEntries[1].fullPath is "/testFiles/file2.txt"
PASS childEntries[1].isFile is true
PASS childEntries[2].name is "subfolder1"
PASS childEntries[2].fullPath is "/testFiles/subfolder1"
PASS childEntries[2].isDirectory is true
PASS childEntries[3].name is "subfolder2"
PASS childEntries[3].fullPath is "/testFiles/subfolder2"
PASS childEntries[3].isDirectory is true
PASS grandChildEntries.length is 1
PASS grandChildEntries[0].name is "file3.txt"
PASS grandChildEntries[0].fullPath is "/testFiles/subfolder1/file3.txt"
PASS grandChildEntries[0].isFile is true
PASS grandChildEntries.length is 2
PASS grandChildEntries[0].name is "file4.txt"
PASS grandChildEntries[0].fullPath is "/testFiles/subfolder2/file4.txt"
PASS grandChildEntries[0].isFile is true
PASS grandChildEntries[1].name is "subfolder2a"
PASS grandChildEntries[1].fullPath is "/testFiles/subfolder2/subfolder2a"
PASS grandChildEntries[1].isDirectory is true
PASS greatGrandChildEntries.length is 1
PASS greatGrandChildEntries[0].name is "file5.txt"
PASS greatGrandChildEntries[0].fullPath is "/testFiles/subfolder2/subfolder2a/file5.txt"
PASS greatGrandChildEntries[0].isFile is true
PASS successfullyParsed is true

TEST COMPLETE
